backend/requirements.txt
django-filter
djangorestframework
djangorestframework-simplejwt
django-cors-headers